+++ src/api-windows.c
@@ -636,3 +636,57 @@ SEXP ps__kill_if_env(SEXP marker, SEXP after, SEXP pid, SEXP sig) {
   UNPROTECT(1);
   return R_NilValue;
 }
+
+SEXP psll_num_fds(SEXP p) {
+  ps_handle_t *handle = R_ExternalPtrAddr(p);
+  HANDLE  hProcess = NULL;
+  DWORD handleCount;
+
+  if (!handle) error("Process pointer cleaned up already");
+
+  hProcess = ps__handle_from_pid(handle->pid);
+  if (hProcess != NULL) {
+    if (GetProcessHandleCount(hProcess, &handleCount)) {
+      CloseHandle(hProcess);
+      PS__CHECK_HANDLE(handle);
+      return ScalarInteger(handleCount);
+    }
+  }
+
+  /* Cleanup on error */
+  if (hProcess != NULL) CloseHandle(hProcess);
+  PS__CHECK_HANDLE(handle);
+  ps__set_error_from_windows_error(0);
+  ps__throw_error();
+  return R_NilValue;
+}
+
+SEXP psll_open_files(SEXP p) {
+  ps_handle_t *handle = R_ExternalPtrAddr(p);
+  HANDLE processHandle = NULL;
+  DWORD access = PROCESS_DUP_HANDLE | PROCESS_QUERY_INFORMATION;
+  SEXP result;
+
+  if (!handle) error("Process pointer cleaned up already");
+
+  processHandle = ps__handle_from_pid_waccess(handle->pid, access);
+  if (processHandle == NULL) {
+    PS__CHECK_HANDLE(handle);
+    ps__set_error_from_windows_error(0);
+    ps__throw_error();
+  }
+
+  PROTECT(result = ps__get_open_files(handle->pid, processHandle));
+
+  CloseHandle(processHandle);
+
+  PS__CHECK_HANDLE(handle);
+
+  if (isNull(result)) {
+    ps__set_error_from_windows_error(0);
+    ps__throw_error();
+  }
+
+  UNPROTECT(1);
+  return result;
+}