+++ src/windows.h
@@ -17,3 +17,4 @@ SEXP ps__proc_resume(DWORD pid);
 SEXP ps__proc_kill(DWORD pid);
 
 double ps__filetime_to_unix(FILETIME ft);
+SEXP ps__convert_dos_path(WCHAR *wstr);