+++ tests/testthat/test-finished.R
@@ -53,4 +53,6 @@ test_that("process already finished", {
   chk(ps_cpu_times(p))
   chk(ps_num_threads(p))
   chk(ps_children(p))
+  chk(ps_num_fds(p))
+  chk(ps_open_files(p))
 })