+++ NEWS.md
@@ -1,4 +1,10 @@
 
+# ps 1.0.0.9000
+
+* `ps_num_fds()` returns the number of open files/handles.
+
+* `ps_open_files()` lists all open files of a process.
+
 # ps 1.0.0
 
 First released version.